requirements.txt
configparser
pyzmq
PySide